backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Security-Too==5.3.2
Flask-CORS==4.0.0
PyMySQL==1.1.0
cryptography==41.0.7
argon2-cffi==23.1.0
email-validator==2.1.0
python-dotenv==1.0.0
mysql-connector-python==8.2.0
reportlab==4.0.4
